memory pool implementation
Summary:
# DIFF 722, control before MemoryBazaar-ized
# DIFF 726, O_LARGEFILE issue and cleaning up MemoryBazaar file added
# DIFF 728, merge with datPRI branch, SIMD ok, but GPU not OK
# DIFF 729, ready for the first read data test
# DIFF 731, solve the bug that THUNDER needs to compile with double make operations
# DIFF 737, CPU version without SIMD done, waiting SIMD modification
# DIFF 743, merge with {D260}, where @ruanhuabin make SIMD work
# DIFF 753, accomplish {T212}
# DIFF 755, fix a bug which contributes to loss of resolution
# DIFF 768, merge with {D256}
# DIFF 795, try to fix a issue raised by @wangzhao
# DIFF 803, {T233} accomplished
# DIFF 808, merge {D256} DIFF 805, merge {D284}
# DIFF 824, merge {D256} DIFF 810

Test Plan: NA

Reviewers: ruanhuabin, #detective_conan

Reviewed By: ruanhuabin, #detective_conan

Subscribers: wangzhao, wangxiaoya, xli

Differential Revision: http://124.65.131.150:11111/D253

+++ include/Functions/Functions.h
@@ -27,17 +27,17 @@
 /**
  * This macros returns the nearest integer number of a.
  */
-#define AROUND(a) ((int)rint(a))
+#define AROUND(a) ((long)rint(a))
 
 /**
  * This macros returns the lower nearest integer number of a.
  */
-#define FLOOR(a) ((int)floor(a))
+#define FLOOR(a) ((long)floor(a))
 
 /**
  * This macros returns the upper nearest integer number of a.
  */
-#define CEIL(a) ((int)ceil(a))
+#define CEIL(a) ((long)ceil(a))
 
 /**
  * This macro returns the maximum value among a and b.
